+++ mongo/client_manager.go
@@ -208,7 +208,7 @@ func (c *ClientManager) Create(ctx context.Context, client storage.Client) (resu
 	collection := c.DB.Collection(storage.EntityClients)
 	_, err = collection.InsertOne(ctx, client)
 	if err != nil {
-		if isDup(err) {
+		if mongo.IsDuplicateKeyError(err) {
 			// Log to StdOut
 			log.WithError(err).Debug(logConflict)
 			// Log to OpenTracing
@@ -398,7 +398,7 @@ func (c *ClientManager) Update(ctx context.Context, clientID string, updatedClie
 	collection := c.DB.Collection(storage.EntityClients)
 	res, err := collection.ReplaceOne(ctx, selector, updatedClient)
 	if err != nil {
-		if isDup(err) {
+		if mongo.IsDuplicateKeyError(err) {
 			// Log to StdOut
 			log.WithError(err).Debug(logConflict)
 			// Log to OpenTracing
@@ -466,7 +466,7 @@ func (c *ClientManager) Migrate(ctx context.Context, migratedClient storage.Clie
 	opts := options.Replace().SetUpsert(true)
 	res, err := collection.ReplaceOne(ctx, selector, migratedClient, opts)
 	if err != nil {
-		if isDup(err) {
+		if mongo.IsDuplicateKeyError(err) {
 			// Log to StdOut
 			log.WithError(err).Debug(logConflict)
 			// Log to OpenTracing