+++ mongo/request_manager.go
@@ -86,6 +86,35 @@ func (r *RequestManager) Configure(ctx context.Context) (err error) {
 	return nil
 }
 
+// ConfigureExpiryWithTTL implements storage.Expirer.
+func (r *RequestManager) ConfigureExpiryWithTTL(ctx context.Context, ttl int) error {
+	collections := []string{
+		storage.EntityAccessTokens,
+		storage.EntityAuthorizationCodes,
+		storage.EntityOpenIDSessions,
+		storage.EntityPKCESessions,
+		storage.EntityRefreshTokens,
+	}
+
+	for _, entityName := range collections {
+		log := logger.WithFields(logrus.Fields{
+			"package":    "mongo",
+			"collection": entityName,
+			"method":     "ConfigureExpiryWithTTL",
+		})
+
+		index := NewExpiryIndex(IdxExpiry+"RequestedAt", "requestedAt", ttl)
+		collection := r.DB.Collection(entityName)
+		_, err := collection.Indexes().CreateOne(ctx, index)
+		if err != nil {
+			log.WithError(err).Error(logError)
+			return err
+		}
+	}
+
+	return nil
+}
+
 // getConcrete returns a Request resource.
 func (r *RequestManager) getConcrete(ctx context.Context, entityName string, requestID string) (result storage.Request, err error) {
 	log := logger.WithFields(logrus.Fields{
@@ -220,7 +249,7 @@ func (r *RequestManager) Create(ctx context.Context, entityName string, request 
 	collection := r.DB.Collection(entityName)
 	_, err = collection.InsertOne(ctx, request)
 	if err != nil {
-		if isDup(err) {
+		if mongo.IsDuplicateKeyError(err) {
 			// Log to StdOut
 			log.WithError(err).Debug(logConflict)
 			// Log to OpenTracing
@@ -317,7 +346,7 @@ func (r *RequestManager) Update(ctx context.Context, entityName string, requestI
 	collection := r.DB.Collection(entityName)
 	res, err := collection.ReplaceOne(ctx, selector, updatedRequest)
 	if err != nil {
-		if isDup(err) {
+		if mongo.IsDuplicateKeyError(err) {
 			// Log to StdOut
 			log.WithError(err).Debug(logConflict)
 			// Log to OpenTracing